+++ postnet/commands/evalInput.js
@@ -1,0 +1,7 @@
+"use strict";
+
+module.exports = function () {
+    return {
+        text:`please give right input:`
+    }
+};
