+++ postnet/commands/exit.js
@@ -1,0 +1,5 @@
+"use strict";
+
+module.exports = function () {
+    process.exit();
+};
