+++ postnet/commands/goToBarcode.js
@@ -1,0 +1,10 @@
+"use strict";
+
+let translateBarCodeToBarcode = require('./translateBarcodeToZipCode');
+
+module.exports = function () {
+    return {
+        text:`please input bar code`,
+        newMapping:{"*":translateBarCodeToBarcode}
+    }
+};
